Refactor Environment models
